reference/netology_pd_diplom/requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==3.6.4.0
celery==5.2.7
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.3
django-rest-passwordreset==1.4.0
djangorestframework==3.14.0
gunicorn==22.0.0
idna==3.6
kombu==5.3.5
packaging==24.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==4.5.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
ujson==5.9.0
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
